chore: refactor folder names
